+++ server/src/main/scala/com/twitter/server/handler/TwitterHandler.scala
@@ -3,10 +3,9 @@ package com.twitter.server.handler
 import com.twitter.finagle.Service
 import com.twitter.finagle.http.{Request, Response, Status}
 import com.twitter.io.Buf
-import com.twitter.server.util.HttpUtils.{newResponse, parse}
+import com.twitter.server.util.HttpUtils.newResponse
 import com.twitter.util.Future
 import com.twitter.util.logging.Logger
-import scala.collection.{Map, Seq}
 
 trait TwitterHandler extends Service[Request, Response] {
   private[this] val log = Logger[TwitterHandler]
@@ -17,9 +16,6 @@ trait TwitterHandler extends Service[Request, Response] {
       status = status,
       content = Buf.Utf8(msg)
     )
-
-  protected[handler] def getParams(uri: String): Map[String, Seq[String]] =
-    parse(uri)._2
 
   protected def log(req: Request, msg: String): Unit = {
     log.info("[%s %s] %s".format(req.method, req.uri, msg))